requirements.txt
fastapi
uvicorn[standard]
openai
langchain
faiss-cpu
python-dotenv
pytest
httpx
loguru
python-multipart
sentence-transformers
torch
prometheus-client
sqlmodel>=0.0.8